1.可以与野火PID调试助手通信

+++ 3.电机程序/1.步进电机_TMC5160/Hardware/MODBUS/modbus.h
@@ -23,7 +23,7 @@ extern uint16_t baudRate_Selection;//波特率选择
 extern void beginSend(void);//开始发送
 extern void b485Send(uint8_t n);//使能485接收发送
 
-void USART1_Init(uint16_t btl);
+void USART1_Init(uint32_t btl);
 void checkComm0Modbus(void);
 void readCoil(void);
 void readRegisters(void);
